requirements.txt
torch==2.0.1
transformers==4.35.0
python-dotenv==1.0.0
psutil==5.9.5